requirements.txt
luma.core==0.1.2
luma.oled==2.0.0
olefile==0.44
Pillow==4.0.0
requests==2.12.4
RPi.GPIO==0.6.3
smbus2==0.1.4
spidev==3.2
